+++ src/Controller/CreateCaseController.php
@@ -5,28 +5,30 @@ namespace App\Controller;
 use App\Entity\ResidentComplaintBoardCase;
 use App\Repository\BoardRepository;
 use App\Repository\MunicipalityRepository;
+use App\Service\CaseManager;
 use Doctrine\DBAL\Exception;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\Routing\Annotation\Route;
 
 class CreateCaseController extends AbstractController
 {
     /**
-     * @Route("/municipality/{municipality_name}/board/{board_name}/case/create", name="rescase")
+     * @Route("/municipality/{municipalityName}/board/{boardName}/case/create", name="rescase")
      */
-    public function createCase(BoardRepository $boardRepository, MunicipalityRepository $municipalityRepository, Request $request, string $municipality_name, string $board_name): Response
+    public function createCase(BoardRepository $boardRepository, CaseManager $caseManager, LockFactory $lockFactory, MunicipalityRepository $municipalityRepository, Request $request, string $municipalityName, string $boardName): Response
     {
         // Check that municipality exists
-        $municipality = $municipalityRepository->findOneBy(['name' => $municipality_name]);
+        $municipality = $municipalityRepository->findOneBy(['name' => $municipalityName]);
 
         if (null === $municipality) {
             throw new Exception('Municipality not found.');
         }
 
         // Check that board exists
-        $board = $boardRepository->findOneBy(['name' => $board_name]);
+        $board = $boardRepository->findOneBy(['name' => $boardName]);
 
         if (null === $board) {
             throw new Exception('Board not found.');
@@ -58,9 +60,20 @@ class CreateCaseController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $case = $form->getData();
 
+            $lock = $lockFactory->createLock('case-number-generation');
+
+            $lock->acquire(true);
+
+            // case number generation
+            $caseNumber = $caseManager->generateCaseNumber($municipality);
+
+            $case->setCaseNumber($caseNumber);
+
             $em = $this->getDoctrine()->getManager();
             $em->persist($case);
             $em->flush();
+
+            $lock->release();
 
             return $this->redirectToRoute('default');
         }